Deprecate moved checksum utils.
