+++ causeway-client/src/main/java/org/jboss/pnc/causewayclient/DefaultCausewayClient.java
@@ -25,12 +25,12 @@ import org.apache.http.message.BasicHeader;
 import org.apache.http.util.EntityUtils;
 import org.jboss.pnc.api.causeway.dto.push.BuildImportRequest;
 import org.jboss.pnc.api.causeway.dto.untag.UntagRequest;
+import org.jboss.pnc.api.constants.MDCKeys;
 import org.jboss.pnc.common.json.GlobalModuleGroup;
 import org.jboss.pnc.common.json.JsonOutputConverterMapper;
 import org.jboss.pnc.common.json.moduleconfig.BpmModuleConfig;
 import org.jboss.pnc.common.logging.MDCUtils;
 import org.jboss.pnc.common.util.HttpUtils;
-import org.jboss.pnc.constants.MDCKeys;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
