+++ common/pom.xml
@@ -78,6 +78,10 @@
     </dependency>
     <dependency>
       <groupId>org.jboss.pnc</groupId>
+      <artifactId>pnc-api</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.jboss.pnc</groupId>
       <artifactId>pnc-common</artifactId>
     </dependency>
     <dependency>
