+++ common/src/main/java/org/jboss/pnc/common/security/CheckSum.java
@@ -24,8 +24,10 @@ import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 
 /**
+ * @deprecated use pnc-common lib
  * @author <a href="mailto:matejonnet@gmail.com">Matej Lazar</a>
  */
+@Deprecated
 public class CheckSum {
     static String calculateDigest(String message, String algorithm) throws NoSuchAlgorithmException, IOException {
         MessageDigest md = MessageDigest.getInstance(algorithm);
