+++ common/src/main/java/org/jboss/pnc/common/security/Sha256.java
@@ -24,8 +24,10 @@ import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 
 /**
+ * @deprecated use pnc-common lib
  * @author <a href="mailto:matejonnet@gmail.com">Matej Lazar</a>
  */
+@Deprecated
 public class Sha256 {
 
     MessageDigest md;
